Removes old join group confirmation dialog
