--- src/app/groups/containers/join-group-confirmation-dialog-v2/join-group-confirmation-dialog.component.html
@@ -1,53 +1,0 @@
-<alg-modal i18n-title title="Joining a new group" (closeEvent)="dialogRef.close()">
-  <p class="description" i18n>Do you want to join the group "{{ data().name }}"?</p>
-  @if (form) {
-    <form class="form" [formGroup]="form">
-      @if (form.get('agreeWithPersonalInfoView')) {
-        <alg-switch-field
-          formControlName="agreeWithPersonalInfoView"
-          data-testid="switch-agree-with-personal-info-view"
-        >
-          <ng-template #label>
-            <span i18n>
-              The managers of this group will be able to
-              { data().params.requirePersonalInfoAccessApproval, select, view {view} other {edit} }
-              the personal info (firstname, lastname, email, grade, ...) you entered in your profile
-            </span>
-          </ng-template>
-        </alg-switch-field>
-      }
-      @if (form.get('agreeWithLockMembership')) {
-        <alg-switch-field
-          class="switch-field"
-          formControlName="agreeWithLockMembership"
-          data-testid="switch-agree-with-lock-membership"
-        >
-          <ng-template #label>
-            <span i18n>
-              You will not be able to leave this group until {{ data().params.requireLockMembershipApprovalUntil | date: 'short' }}
-            </span>
-          </ng-template>
-        </alg-switch-field>
-      }
-    </form>
-  }
-
-  <footer>
-    <div class="footer-container">
-      <button
-        class="stroke button"
-        icon="ph-bold ph-x"
-        (click)="dialogRef.close()"
-        alg-button
-        i18n
-      >Cancel</button>
-      <button
-        icon="ph-bold ph-check"
-        (click)="dialogRef.close({ confirmed: true })"
-        [disabled]="form?.invalid"
-        alg-button
-        i18n
-      >Join group</button>
-    </div>
-  </footer>
-</alg-modal>
