--- src/app/groups/containers/join-group-confirmation-dialog-v2/join-group-confirmation-dialog.component.scss
@@ -1,28 +1,0 @@
-@import '../../../../assets/scss/functions';
-
-:host {
-  display: block;
-  width: 50vw;
-}
-
-.button {
-  margin-right: var(--alg-space-size-6);
-}
-
-.description {
-  margin: 0;
-}
-
-.form {
-  margin-top: toRem(16);
-}
-
-.switch-field {
-  display: block;
-  margin-top: var(--alg-space-size-5);
-}
-
-.footer-container {
-  display: flex;
-  justify-content: center;
-}
