+++ src/app/groups/containers/join-group-confirmation-dialog/join-group-confirmation-dialog.component.ts
@@ -1,11 +1,13 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { DialogModule } from 'primeng/dialog';
+import { Component, inject, OnInit, signal } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { SwitchFieldComponent } from 'src/app/ui-components/collapsible-section/switch-field/switch-field.component';
 import { DatePipe } from '@angular/common';
 import { GroupApprovals, RequirePersonalInfoAccessApproval as PIApproval } from 'src/app/groups/models/group-approvals';
 import { ButtonComponent } from 'src/app/ui-components/button/button.component';
-import { ButtonIconComponent } from 'src/app/ui-components/button-icon/button-icon.component';
+import { ModalComponent } from 'src/app/ui-components/modal/modal.component';
+import { DIALOG_DATA, DialogRef } from '@angular/cdk/dialog';
+
+export type JoinGroupConfirmationDialogResult = { confirmed: true } | undefined;
 
 @Component({
   selector: 'alg-join-group-confirmation-dialog',
@@ -13,19 +15,17 @@ import { ButtonIconComponent } from 'src/app/ui-components/button-icon/button-ic
   styleUrls: [ './join-group-confirmation-dialog.component.scss' ],
   standalone: true,
   imports: [
-    DialogModule,
     SwitchFieldComponent,
     ReactiveFormsModule,
     DatePipe,
     ButtonComponent,
-    ButtonIconComponent,
+    ModalComponent,
   ],
 })
 export class JoinGroupConfirmationDialogComponent implements OnInit {
-  @Output() cancelEvent = new EventEmitter<void>();
-  @Output() confirmEvent = new EventEmitter<void>();
-  @Input() name = '';
-  @Input({ required: true }) params!: GroupApprovals;
+  dialogRef = inject(DialogRef<JoinGroupConfirmationDialogResult>);
+
+  data = signal(inject<{ name: string, params: GroupApprovals }>(DIALOG_DATA));
 
   form?: FormGroup<{
     agreeWithPersonalInfoView?: FormControl<boolean>,
@@ -36,7 +36,7 @@ export class JoinGroupConfirmationDialogComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    const { requirePersonalInfoAccessApproval, requireLockMembershipApprovalUntil } = this.params;
+    const { requirePersonalInfoAccessApproval, requireLockMembershipApprovalUntil } = this.data().params;
     if (requirePersonalInfoAccessApproval !== PIApproval.None || requireLockMembershipApprovalUntil) {
       this.form = this.fb.nonNullable.group({});
       if (requirePersonalInfoAccessApproval !== PIApproval.None) {
